split build selector panels into separate files
